avoid collision with coldfront allocation url
The Django REST framework router autogenerates URL names based on the
queryset if no basename is specified. In the case of allocations, this
results in `allocation-list` and `allocation-detail` url names being
generated which conflicts with the upstream allocation urls/view names.
Adding an explicit basename fixes this by using a different prefix