Fix wrong env var check for OIDC
The plugin expects `PLUGIN_OIDC` for OpenID Connect when it should
be `PLUGIN_AUTH_OIDC`.

This causes the plugin to not correctly configure OIDC for
authentication of the `/api/allocations/` endpoint and defaults to
BasicAuthentication and SessionAuthentication.

Additionally, this keeps SessionAuthentiation because it's useful
for browsing the API through a web browser.