Allow adding of users to groups from searching IdP
ColdFront allows to configure external search providers, if a
user doesn't exist in the ColdFront database. This is what powers
the search function in the web browser version.

This extends that functionality to the API.

You can now add a user to a group, if the user doesn't exist in
ColdFront, but instead exists in one of the configured search
providers, in our case that would be Keycloak.

Additionally, this adds tests and mock and helper modules to
keep the code tidy.