Allow preregistering users without an identity provider account
This is necessary to pre-register users that haven't yet registered
with the identity provider, for the NERC through RegApp for example.

Since they don't have a corresponding IdP account, they won't be able
to log in yet, but this unblocks the PIs and cloud administrators from
having to register all users with the identity provider, before they
can add them to projects.