Implemented management command to fix database inconsistency (#18)
I implemented a management command named fix_allocation_users that adds users to the projects that their assigned allocations are a part of. Running the command by itself only logs out which user has this database inconsistency problem, and using --apply fixes those inconsistencies.

The test_fix_allocation test is added to test for this command.