Added functional testing for SCIM authentication
A new CI job has been added to test the OIDC authentication for the
API endpoints. This CI starts Keycloak in a container, installs coldfront
then runs the API functional tests.

A few code decisions to note:
- The functional tests subclassed the keycloak client provided by `CCI-MOC/onboarding-tools`.
Subclassing was needed because the original client class contained code which is out-of-date
with the current version of Keycloak
- The `is_user_superuser()` function in `utils.py` now also checks if `user.is_superuser`,
to account for edge case where a superuser is created but their `is_staff` attribute is
not set.