Resolves CI failure due to bad patch failure
The latest version of `coldfront` (v1.1.7) made a change to
`coldfront/config/urls.py b/coldfront/config/urls.py`, which caused
`01_add_coldfront_urls.patch` to be out-of-date. This meant the
CI script was unable to apply the patch.

A version pin for `coldfront` has been added to `requirements.txt`,
matching the version used in `coldfront-nerc` (v1.1.6).

The order of pip install commands in `ci/setup.sh` has been changed
to resolve a much more convoluted dependancy-related issue as described below.

`coldfront v1.1.6` required `django-q` [1],
while `coldfront v1.1.7` required `django-q2` [2].
Both packages provide the `django_q` module, and are
incompatible with each other.
The `test-requirements.txt` file in this repo requires
`coldfront-plugin-cloud`, which doesn't set a upper limit
on the version of `coldfront`.
`coldfront-plugin-api` requires `coldfront v1.1.6`.
This means when `test-requirements.txt` is installed first, it will
install `coldfront v1.1.7` and `django-q2`. Afterwards, when
`coldfront-plugin-api` is installed, it will downgrade `coldfront`
to v1.1.6, and install `django-q`, which will corrupt the `django_q` module.
By reversing the order of installation, `coldfront` is
never downgraded, avoiding the corruption.
All of this stems from the fact that `coldfront-plugin-cloud`
doesn't set an upper limit on the version of `coldfront` it supports.

[1] https://github.com/Koed00/django-q/tree/master
[2] https://github.com/django-q2/django-q2